mcmctree

+++ PhyloSuite/src/Lg_MCMCTree.py
@@ -1,5 +1,8 @@
+import datetime
 import multiprocessing
 import re
+import shutil
+import traceback
 
 from PyQt5.QtGui import *
 from PyQt5.QtCore import *
@@ -16,11 +19,17 @@ from uifiles.Ui_mcmctree import Ui_MCMCTree
 class MCMCTree(QDialog,Ui_MCMCTree,object):
     showSig = pyqtSignal(QDialog)
     closeSig = pyqtSignal(str, str)
+    logGuiSig = pyqtSignal(str)
+    startButtonStatusSig = pyqtSignal(list)
+    mcmctree_exception = pyqtSignal(str)
+    exception_signal = pyqtSignal(str)  # 定义所有类都可以使用的信号
+    progressSig = pyqtSignal(int)  # 控制进度条
 
     def __init__(
             self,
             workPath=None,
             focusSig=None,
+            mcmctreeEXE=None,
             parent=None):
         super(MCMCTree, self).__init__(parent)
         self.parent = parent
@@ -29,6 +38,7 @@ class MCMCTree(QDialog,Ui_MCMCTree,object):
         self.thisPath = self.factory.thisPath
         self.workPath = workPath
         self.focusSig = focusSig
+        self.mcmctreeEXE = mcmctreeEXE
         self.seqFileName = ''
         self.treeFileName = ''
         self.setupUi(self)
@@ -36,16 +46,27 @@ class MCMCTree(QDialog,Ui_MCMCTree,object):
            self.thisPath + '/settings/MCMCTree_settings.ini', QSettings.IniFormat)
         # File only, no fallback to registry or or.
         self.MCMCTree_settings.setFallbacksEnabled(False)
-        # self.dict_tax_color = {}
-        # 开始装载样式表
-        # with open(self.thisPath + os.sep + 'style.qss', encoding="utf-8", errors='ignore') as f:
-        #     self.qss_file = f.read()
-        # self.setStyleSheet(self.qss_file)
         self.qss_file = self.factory.set_qss(self)
         # 恢复用户的设置
         self.guiRestore()
         self.log_gui = self.gui4Log()
         self.text_gui = self.gui4Text()
+        # 槽函数
+        self.startButtonStatusSig.connect(self.factory.ctrl_startButton_status)
+        self.logGuiSig.connect(self.addText2Log)
+        self.progressSig.connect(self.runProgress)
+        # 给开始按钮添加菜单
+        menu = QMenu(self)
+        menu.setToolTipsVisible(True)
+        self.work_action = QAction(QIcon(":/picture/resourses/work.png"), "", menu)
+        self.work_action.triggered.connect(lambda: self.factory.swithWorkPath(self.work_action, parent=self))
+        self.dir_action = QAction(QIcon(":/picture/resourses/folder.png"), "Output Dir: ", menu)
+        self.dir_action.triggered.connect(lambda: self.factory.set_direct_dir(self.dir_action, self))
+        menu.addAction(self.work_action)
+        menu.addAction(self.dir_action)
+        self.pushButton_5.toolButton.setMenu(menu)
+        self.pushButton_5.toolButton.menu().installEventFilter(self)
+        self.factory.swithWorkPath(self.work_action, init=True, parent=self)  # 初始化一下
 
     def input(self, file, which):
         base = os.path.basename(file)
@@ -76,7 +97,7 @@ class MCMCTree(QDialog,Ui_MCMCTree,object):
             self.input(self.treeFileName, 3)
 
     @pyqtSlot()
-    def on_pushButton_5_clicked(self):
+    def on_pushButton_15_clicked(self):
         s_Values, ds_Values = self.getParas()
         print(f"s_Values:{s_Values}")
         print(f"ds_Values:{ds_Values}")
@@ -92,6 +113,122 @@ class MCMCTree(QDialog,Ui_MCMCTree,object):
     @pyqtSlot()
     def on_pushButton_7_clicked(self):
         self.text_gui.show()
+
+    @pyqtSlot()
+    def on_pushButton_5_clicked(self):
+        """
+        execute program
+        """
+        self.commands = self.prepare_for_run()
+        print(self.commands)
+        if self.commands:
+            # 有数据才执行
+            self.interrupt = False  # 如果是终止以后重新运行要刷新
+            self.textEdit_log.clear()  # 清空log
+            self.mcmctree_popen = self.factory.init_popen(self.commands)
+            self.factory.emitCommands(self.logGuiSig, self.commands)
+            self.worker = WorkThread(self.run_command, parent=self)
+            self.worker.start()
+
+    def run_command(self):
+        try:
+            time_start = datetime.datetime.now()
+            self.startButtonStatusSig.emit(
+                [
+                    self.pushButton_5,
+                    self.progressBar,
+                    "start",
+                    self.exportPath,
+                    self.qss_file,
+                    self])
+            print(111)
+            self.run_code()
+            time_end = datetime.datetime.now()
+            self.time_used = str(time_end - time_start)
+            self.time_used_des = "Start at: %s\nFinish at: %s\nTotal time used: %s\n\n" % (str(time_start), str(time_end),
+                                                                                           self.time_used)
+            # 加上version信息
+            with open(self.exportPath + os.sep + "PhyloSuite_PartitionFinder.log") as f:
+                content = f.read()
+            rgx_version = re.compile(r"-+\s+PartitionFinder.+?(\d+\.\d+\.\d+)")
+            if rgx_version.search(content):
+                self.version = rgx_version.search(content).group(1)
+            else:
+                self.version = ""
+            self.description = self.description.replace("$version$", self.version)
+            with open(self.exportPath + os.sep + "summary and citation.txt", "w", encoding="utf-8") as f:
+                f.write(self.description +
+                        f"\n\nIf you use PhyloSuite v1.2.3 v1.2.3, please cite:\n{self.factory.get_PS_citation()}\n\n"
+                        "If you use PartitionFinder 2, please cite:\n" + self.reference + "\n\n" + self.time_used_des)
+            if not self.interrupt:
+                if self.workflow:
+                    # work flow跑的
+                    self.startButtonStatusSig.emit(
+                        [
+                            self.pushButton_5,
+                            self.progressBar,
+                            "workflow stop",
+                            self.exportPath,
+                            self.qss_file,
+                            self])
+                    self.workflow_finished.emit("finished")
+                    return
+                self.startButtonStatusSig.emit(
+                    [
+                        self.pushButton_5,
+                        self.progressBar,
+                        "stop",
+                        self.exportPath,
+                        self.qss_file,
+                        self])
+            else:
+                self.startButtonStatusSig.emit(
+                    [
+                        self.pushButton_5,
+                        self.progressBar,
+                        "except",
+                        self.exportPath,
+                        self.qss_file,
+                        self])
+            if not self.workflow:
+                self.focusSig.emit(self.exportPath)
+        except BaseException:
+            self.exceptionInfo = ''.join(
+                traceback.format_exception(
+                    *sys.exc_info()))  # 捕获报错内容，只能在这里捕获，没有报错的地方无法捕获
+            self.exception_signal.emit(self.exceptionInfo)  # 激发这个信号
+            self.startButtonStatusSig.emit(
+                [
+                    self.pushButton_5,
+                    self.progressBar,
+                    "except",
+                    self.exportPath,
+                    self.qss_file,
+                    self])
+
+    def run_code(self):
+        is_error = False  ##判断是否出了error
+        error_text = ""
+        self.progressSig.emit(99999)
+        while True:
+            QApplication.processEvents()
+            if self.isRunning():
+                try:
+                    out_line = self.mcmctree_popen.stdout.readline().decode("utf-8", errors="ignore")
+                except UnicodeDecodeError:
+                    out_line = self.mcmctree_popen.stdout.readline().decode("gbk", errors="ignore")
+                if out_line == "" and self.mcmctree_popen.poll() is not None:
+                    break
+                self.logGuiSig.emit(out_line.strip())
+                if re.search(r"^ERROR", out_line):
+                    is_error = True
+                    error_text = "Error happened! Click <span style='font-weight:600; color:#ff0000;'>Show log</span> to see detail!"
+            else:
+                break
+        if is_error:
+            self.interrupt = True
+            self.mcmctree_exception.emit(error_text)
+        self.mcmctree_popen = None
 
     def showText(self, lst):
         self.tableView_2.setRowCount(len(lst))
@@ -324,15 +461,9 @@ class MCMCTree(QDialog,Ui_MCMCTree,object):
                 ds_Values.append(obj.value())
         return s_Values, ds_Values
 
-    def ctl_generater(self,
-                      ):
-        s_Values, ds_Values = self.getParas()
-        options = QFileDialog.DontResolveSymlinks | QFileDialog.ShowDirsOnly
-        directory = QFileDialog.getExistingDirectory(self, "Choose folder", options=options)
-        if directory:
-            seqfile = os.path.basename(self.seqFileName)
-            treefile = os.path.basename(self.treeFileName)
-            list_mcmctree_ctl = [f'''          seed = {s_Values[15]}
+    def ctl_generater(self, file_path=None):
+        def generate_ctl(seqfile, treefile):
+            return f'''          seed = {s_Values[15]}
        seqfile = {seqfile}
       treefile = {treefile}
       mcmcfile = mcmc.txt
@@ -362,17 +493,75 @@ class MCMCTree(QDialog,Ui_MCMCTree,object):
       sampfreq = {s_Values[10]}
        nsample = {s_Values[12]}
 
-*** Note: Make your window wider (100 columns) before running the program.''']
-            file_path = f"{directory}{os.sep}mcmctree.ctl"
-            if os.path.exists(file_path):
-                reply = QMessageBox.question(self, "File Exists", f"文件 {file_path} 已存在，是否覆盖？",
-                                             QMessageBox.Yes | QMessageBox.No)
-                if reply == QMessageBox.Yes:
+*** Note: Make your window wider (100 columns) before running the program.'''
+        s_Values, ds_Values = self.getParas()
+        seqfile = os.path.basename(self.seqFileName)
+        treefile = os.path.basename(self.treeFileName)
+        if file_path:
+            mcmctree_ctl = generate_ctl(seqfile, treefile)
+            with open(file_path, "w", errors="ignore") as f:
+                f.write(mcmctree_ctl)
+        else:
+            options = QFileDialog.DontResolveSymlinks | QFileDialog.ShowDirsOnly
+            directory = QFileDialog.getExistingDirectory(self, "Choose folder", options=options)
+            if directory:
+                mcmctree_ctl = generate_ctl(seqfile, treefile)
+                file_path = f"{directory}{os.sep}mcmctree.ctl"
+                if os.path.exists(file_path):
+                    reply = QMessageBox.question(self, "File Exists", f"文件 {file_path} 已存在，是否覆盖？",
+                                                 QMessageBox.Yes | QMessageBox.No)
+                    if reply == QMessageBox.Yes:
+                        with open(file_path, "w", errors="ignore") as f:
+                            f.write(mcmctree_ctl)
+                else:
                     with open(file_path, "w", errors="ignore") as f:
-                        f.write("\n".join(list_mcmctree_ctl))
-            else:
-                with open(file_path, "w", errors="ignore") as f:
-                    f.write("\n".join(list_mcmctree_ctl))
+                        f.write(mcmctree_ctl)
+
+    def prepare_for_run(self):
+        self.output_dir_name = self.factory.fetch_output_dir_name(self.dir_action)
+        self.exportPath = self.factory.creat_dirs(self.workPath +
+                                                  os.sep + "MCMCTREE_results" + os.sep + self.output_dir_name)
+        ctl_file = f"{self.exportPath}{os.sep}mcmctree.ctl"
+        self.ctl_generater(ctl_file)
+        shutil.copy(self.seqFileName, self.exportPath + os.sep + os.path.basename(self.seqFileName))
+        shutil.copy(self.treeFileName, self.exportPath + os.sep + os.path.basename(self.treeFileName))
+        os.chdir(self.exportPath)
+        cmds = f"{self.mcmctreeEXE} {ctl_file}"
+        return cmds
+
+
+    def eventFilter(self, obj, event):
+        # modifiers = QApplication.keyboardModifiers()
+        name = obj.objectName()
+        if (event.type() == QEvent.Show) and (obj == self.pushButton_5.toolButton.menu()):
+            if re.search(r"\d+_\d+_\d+\-\d+_\d+_\d+",
+                         self.dir_action.text()) or self.dir_action.text() == "Output Dir: ":
+                self.factory.sync_dir(self.dir_action)  ##同步文件夹名字
+            menu_x_pos = self.pushButton_5.toolButton.menu().pos().x()
+            menu_width = self.pushButton_5.toolButton.menu().size().width()
+            button_width = self.pushButton_5.toolButton.size().width()
+            pos = QPoint(menu_x_pos - menu_width + button_width,
+                         self.pushButton_5.toolButton.menu().pos().y())
+            self.pushButton_5.toolButton.menu().move(pos)
+            return True
+        # return QMainWindow.eventFilter(self, obj, event) #
+        # 其他情况会返回系统默认的事件处理方法。
+        return super(MCMCTree, self).eventFilter(obj, event)  # 0
+
+    def isRunning(self):
+        '''判断程序是否运行,依赖进程是否存在来判断'''
+        return hasattr(self, "mcmctree_popen") and self.mcmctree_popen and not self.interrupt
+
+    def addText2Log(self, text):
+        if re.search(r"\w+", text):
+            self.textEdit_log.append(text)
+            with open(self.exportPath + os.sep + "PhyloSuite_MCMCTREE.log", "a", errors='ignore') as f:
+                f.write(text + "\n")
+
+    def runProgress(self, num):
+        if num == 99999:
+            self.progressBar.setMaximum(0)
+            self.progressBar.setMinimum(0)
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
